% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_Threshold.R
\name{Param_thresh}
\alias{Param_thresh}
\title{Estimates the threshold function E_W E[Y | A >=c, W] for a range of given values threshold values c}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Estimates the threshold function E_W E[Y | A >=c, W] for a range of given values threshold values c

Estimates the threshold function E_W E[Y | A >=c, W] for a range of given values threshold values c
}
\section{Constructor}{

  \code{define_param(Param_ATT, observed_likelihood, intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention.
    }
    \item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention.
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood_treatment}}{the counterfactual likelihood for the treatment
    }
    \item{\code{cf_likelihood_control}}{the counterfactual likelihood for the control
    }
    \item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention
    }
    \item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention
    }
}
}

\concept{Parameters}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:Param_base]{tmle3::Param_base}} -> \code{Param_thresh}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_thresh$new()}}
\item \href{#method-clever_covariates}{\code{Param_thresh$clever_covariates()}}
\item \href{#method-estimates}{\code{Param_thresh$estimates()}}
\item \href{#method-clone}{\code{Param_thresh$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Param_base" data-id="print">}\href{../../tmle3/html/Param_base.html#method-print}{\code{tmle3::Param_base$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Param_base" data-id="supports_submodel_type">}\href{../../tmle3/html/Param_base.html#method-supports_submodel_type}{\code{tmle3::Param_base$supports_submodel_type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_thresh$new(
  observed_likelihood,
  thresh_node = "A",
  outcome_node = "Y",
  type = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates"></a>}}
\if{latex}{\out{\hypertarget{method-clever_covariates}{}}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_thresh$clever_covariates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimates"></a>}}
\if{latex}{\out{\hypertarget{method-estimates}{}}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_thresh$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_thresh$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
