% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_spec_Threshold.R
\name{tmle3_Spec_Threshold}
\alias{tmle3_Spec_Threshold}
\title{Defines a TML Estimator (except for the data)}
\description{
Defines a TML Estimator (except for the data)

Defines a TML Estimator (except for the data)
}
\section{Super class}{
\code{\link[tmle3:tmle3_Spec]{tmle3::tmle3_Spec}} -> \code{tmle3_Spec_Threshold}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tmle3_Spec_Threshold$new()}}
\item \href{#method-make_tmle_task}{\code{tmle3_Spec_Threshold$make_tmle_task()}}
\item \href{#method-make_initial_likelihood}{\code{tmle3_Spec_Threshold$make_initial_likelihood()}}
\item \href{#method-make_params}{\code{tmle3_Spec_Threshold$make_params()}}
\item \href{#method-clone}{\code{tmle3_Spec_Threshold$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_targeted_likelihood">}\href{../../tmle3/html/tmle3_Spec.html#method-make_targeted_likelihood}{\code{tmle3::tmle3_Spec$make_targeted_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_updater">}\href{../../tmle3/html/tmle3_Spec.html#method-make_updater}{\code{tmle3::tmle3_Spec$make_updater()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_Threshold$new(
  threshold_function = function(A) {     as.vector(quantile(A, seq(0.05, 0.95,
    length.out = 10))) },
  cdf_bins = 10,
  data_adaptive = F,
  cv = data_adaptive,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_tmle_task"></a>}}
\if{latex}{\out{\hypertarget{method-make_tmle_task}{}}}
\subsection{Method \code{make_tmle_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_Threshold$make_tmle_task(data, node_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_initial_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-make_initial_likelihood}{}}}
\subsection{Method \code{make_initial_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_Threshold$make_initial_likelihood(tmle_task, learner_list = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_params"></a>}}
\if{latex}{\out{\hypertarget{method-make_params}{}}}
\subsection{Method \code{make_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_Threshold$make_params(tmle_task, likelihood)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_Threshold$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
