% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_spec_Threshold.R
\name{tmle_Threshold}
\alias{tmle_Threshold}
\title{All Treatment Specific Means}
\usage{
tmle_Threshold(
  threshold_function = function(A) {     as.vector(quantile(A, seq(0.05, 0.95,
    length.out = 10))) },
  cdf_bins = 10,
  data_adaptive = F,
  cv = data_adaptive,
  ...
)
}
\arguments{
\item{treatment_level}{the level of A that corresponds to treatment}

\item{control_level}{the level of A that corresponds to a control or reference level}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
